Attach event listener to form submit button
Once the user fills out everything in the form
and clicks on the submit button. Then the event
listener callback is triggered in which each
individual input will be stored in four variables,
which are title, author, pages, and read. If the
user does not click on the checkbox within read,
then the value will be an empty string.