Add styling for user book entry
The cards container will contain 3 book entries
in one row. If the user inputs more than 3, the
subsequent book will be wrapped to a new row.
To put it simply, the children of the cards
container, aka the book entry will have an
explicit height. For now, the styling of the
border and the border-radius will remain intact.