Attach event listener to read button
Upon creation and attachment of book to
library, an event listener for the read
button is attached. If the user clicks
on the read button, then the button
switches to either Not Read or Read.
Additionally, the read status of the
current book is also updated within
the array referenced by the variable
myLibrary.